requirements.txt
Django==5.2.3
django-rest-framework==0.1.0
django-filter==25.1
celery==5.5.3
django-celery-beat==2.8.1
psycopg2-binary>=2.9
redis==6.2.0